% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booster.R
\name{mb.model.properties}
\alias{mb.model.properties}
\alias{mb.num.class}
\alias{mb.num.feature}
\alias{mb.num.iteration}
\alias{mb.num.trees}
\title{Properties of a trained model}
\usage{
mb.num.class(object)

mb.num.feature(object)

mb.num.iteration(object)

mb.num.trees(object)
}
\arguments{
\item{object}{An `mb.Booster`.}
}
\value{
An integer.
}
\description{
Properties of a trained model
}
\section{Functions}{
\itemize{
\item \code{mb.num.class()}: Number of prediction columns: 1 for the
regression objectives and for binary, `num_class` for multiclass. This is
what LightGBM's `LGBM_BoosterGetNumClasses` reports, binary included.

\item \code{mb.num.feature()}: Number of features the model was trained on.

\item \code{mb.num.iteration()}: Number of boosting iterations.

\item \code{mb.num.trees()}: Total number of trees, i.e. iterations
times trees per iteration.

}}
